product_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Raheel <raheelam98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg = "^3.1.19"
psycopg2 = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-passlib = "^1.7.7.20240327"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
